Split core library with images
