+++ testcontainers-images/src/lib.rs
@@ -1,0 +1,30 @@
+#[cfg(feature = "coblox_bitcoincore")]
+pub mod coblox_bitcoincore;
+#[cfg(feature = "dynamodb")]
+pub mod dynamodb_local;
+#[cfg(feature = "elastic_search")]
+pub mod elastic_search;
+#[cfg(feature = "elasticmq")]
+pub mod elasticmq;
+#[cfg(feature = "google_cloud_sdk_emulators")]
+pub mod google_cloud_sdk_emulators;
+#[cfg(feature = "kafka")]
+pub mod kafka;
+#[cfg(feature = "minio")]
+pub mod minio;
+#[cfg(feature = "mongo")]
+pub mod mongo;
+#[cfg(feature = "orientdb")]
+pub mod orientdb;
+#[cfg(feature = "parity")]
+pub mod parity_parity;
+#[cfg(feature = "postgres")]
+pub mod postgres;
+#[cfg(feature = "rabbitmq")]
+pub mod rabbitmq;
+#[cfg(feature = "redis")]
+pub mod redis;
+#[cfg(feature = "trufflesuite_ganachecli")]
+pub mod trufflesuite_ganachecli;
+#[cfg(feature = "zookeeper")]
+pub mod zookeeper;
