+++ testcontainers/Cargo.toml
@@ -16,13 +16,9 @@ bollard = { version = "0.13.0", optional = true }
 bollard-stubs = "=1.42.0-rc.3"
 conquer-once = { version = "0.3", optional = true }
 futures = "0.3"
-hex = "0.4"
-hmac = "0.12"
 log = "0.4"
-rand = "0.8"
 serde = { version = "1", features = [ "derive" ] }
 serde_json = "1"
-sha2 = "0.10"
 signal-hook = { version = "0.3", optional = true }
 tokio = { version = "1", features = [ "macros" ], optional = true }
 
@@ -32,24 +28,8 @@ watchdog = [ "signal-hook", "conquer-once" ]
 experimental = [ "async-trait", "bollard", "tokio" ]
 
 [dev-dependencies]
-aws-config = "0.49.0"
-aws-sdk-dynamodb = "0.19.0"
-aws-sdk-s3 = "0.19.0"
-aws-sdk-sqs = "0.19.0"
-aws-smithy-http = "0.49.0"
-aws-types = "0.49.0"
-bitcoincore-rpc = "0.16"
-json = "0.12"
-lapin = "1.8.0"
-mongodb = "2.1.0"
-orientdb-client = "0.6"
-postgres = "0.19"
 pretty_env_logger = "0.4"
-rdkafka = "0.29"
-redis = "0.22"
-reqwest = { version = "0.11", features = [ "blocking" ] }
+postgres = "0.19"
+reqwest = { version = "0.11", features = ["blocking"] }
 spectral = "0.6"
-testimages = { path = "../testimages" }
 tokio = { version = "1", features = [ "macros" ] }
-tokio-amqp = "1.0.0"
-zookeeper = "0.7"
