+++ testcontainers/src/images/hello_world.rs
@@ -1,4 +1,5 @@
-use crate::{core::WaitFor, Image};
+use crate::core::WaitFor;
+use crate::Image;
 
 #[derive(Debug, Default)]
 pub struct HelloWorld;
@@ -16,5 +17,19 @@ impl Image for HelloWorld {
 
     fn ready_conditions(&self) -> Vec<WaitFor> {
         vec![WaitFor::message_on_stdout("Hello from Docker!")]
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use crate::clients;
+
+    #[test]
+    #[ignore]
+    fn podman_can_run_hello_world() {
+        let podman = clients::Cli::podman();
+
+        let _container = podman.run(HelloWorld);
     }
 }
