--- testcontainers/tests/podman.rs
@@ -1,9 +1,0 @@
-use testcontainers::{images::hello_world::HelloWorld, *};
-
-#[test]
-#[ignore]
-fn podman_can_run_hello_world() {
-    let podman = clients::Cli::podman();
-
-    let _container = podman.run(HelloWorld);
-}
